Renaming some project name to temporary fix a repo command bug
On repo (or git-repo), If you use the same project name with another
one, but sync them in different paths, you can't be able to sync the
submodules if the submodules are the same on both paths. We've found
this bug while syncing KernelSU for the generic kernel and the surface
kernel. To temporary fix this, we set one character in these names to
CAPS, making repo recognize them differently thus syncing our submodules.

Here's the issue page that is related to this bug, we will revert once it's
being fixed :
https://issues.gerritcodereview.com/issues/40013218

Change-Id: I3228e6eda2ba4e4fb07ea1f01b006a59dd9267df
Signed-off-by: hmtheboy154 <buingoc67@gmail.com>